+++ packages/animated-pose/src/factory.ts
@@ -174,14 +174,10 @@ export default ({
      * Return the Poser API returned by the factory function, with extra methods
      * specific to Animated Pose
      */
-    extendAPI: api => {
-      return {
-        ...api,
-        addChild: (props: AnimatedPoseConfig) => {
-          return api._addChild(props, pose);
-        }
-      };
-    }
+    extendAPI: api => ({
+      ...api,
+      addChild: (config: AnimatedPoseConfig) => api._addChild(config, pose)
+    })
   });
 
   return pose;