+++ packages/popmotion-pose/src/dom/flip.ts
@@ -1,27 +1,19 @@
 import value from 'popmotion/reactions/value';
 import {
-  Pose,
-  ValueMap,
-  PoseSetterFactoryProps,
   BoundingBox,
-  Dimensions
+  Dimensions,
+  Value,
+  Pose,
+  PoserState,
+  DomPopmotionPoser
 } from '../types';
+import { Poser } from 'pose-core';
 import { Action } from 'popmotion/action';
-import { Styler } from 'stylefire';
-
-// Prevents the bug where TS errors between "export cannot be named"
-// and import is "declared but unused".
-export { Dimensions };
-export { Action };
+import { ColdSubscription } from 'popmotion/action/types';
 
 const ORIGIN_START = 0;
 const ORIGIN_CENTER = '50%';
 const ORIGIN_END = '100%';
-
-type SetValueProps = {
-  values: ValueMap;
-  elementStyler: Styler;
-};
 
 type FlipPose = {
   scaleX?: number;
@@ -30,30 +22,27 @@ type FlipPose = {
   y?: number;
 };
 
+type StyleMap = { [key: string]: any };
+
 const findCenter = ({ top, right, bottom, left }: BoundingBox) => ({
   x: (left + right) / 2,
   y: (top + bottom) / 2
 });
 
-const positionalProps = new Set([
-  'width',
-  'height',
-  'top',
-  'left',
-  'bottom',
-  'right'
-]);
-const checkPositionalProp = (key: string) => positionalProps.has(key);
+const positionalProps = ['width', 'height', 'top', 'left', 'bottom', 'right'];
+const positionalPropsDict = new Set(positionalProps);
+const checkPositionalProp = (key: string) => positionalPropsDict.has(key);
 const hasPositionalProps = (pose: Pose) =>
   Object.keys(pose).some(checkPositionalProp);
-export const isFlipPose = (pose: Pose, key: string) =>
-  hasPositionalProps(pose) || key === 'flip';
 
-const setValue = (
-  { values, elementStyler }: SetValueProps,
-  key: string,
-  to: any
-) => {
+export const isFlipPose = (pose: Pose, key: string, state: PoserState) =>
+  state.props.element instanceof HTMLElement &&
+  (hasPositionalProps(pose) || key === 'flip');
+
+export const resolveProp = (target: any, props: { [key: string]: any }) =>
+  typeof target === 'function' ? target(props) : target;
+
+const setValue = ({ values, props }: PoserState, key: string, to: any) => {
   if (values.has(key)) {
     // Here, if we already have the value, we update it twice.
     // Because of stylefire's render batching, this isn't going
@@ -61,16 +50,19 @@ const setValue = (
     // the value jump a great distance, we want to reset the velocity
     // to 0, rather than something arbitrarily high
     // Maybe a more explicit API would be nicer
-    const val = values.get(key);
-    val.update(to);
-    val.update(to);
+    const { raw } = values.get(key);
+    raw.update(to);
+    raw.update(to);
   } else {
-    values.set(key, value(to, (v: any) => elementStyler.set(key, v)));
+    values.set(key, {
+      raw: value(to, (v: any) => props.elementStyler.set(key, v))
+    });
   }
 };
 
-const explicitlyFlipPose = (state: PoseSetterFactoryProps, nextPose: Pose) => {
-  const { dimensions, elementStyler } = state;
+const explicitlyFlipPose = (state: PoserState, nextPose: Pose) => {
+  const { dimensions, elementStyler } = state.props;
+
   dimensions.measure();
   const {
     width,
@@ -82,20 +74,25 @@ const explicitlyFlipPose = (state: PoseSetterFactoryProps, nextPose: Pose) => {
     ...remainingPose
   } = nextPose;
 
-  (elementStyler.set({
-    width,
-    height,
-    top,
-    left,
-    bottom,
-    right
-  }) as Styler).render();
+  elementStyler
+    .set(
+      positionalProps.reduce(
+        (acc, key) => {
+          if (nextPose[key] !== undefined) {
+            acc[key] = resolveProp(nextPose[key], state.props);
+          }
+          return acc;
+        },
+        {} as StyleMap
+      )
+    )
+    .render();
 
   return implicitlyFlipPose(state, remainingPose);
 };
 
-const implicitlyFlipPose = (state: PoseSetterFactoryProps, nextPose: Pose) => {
-  const { dimensions, element, elementStyler } = state;
+const implicitlyFlipPose = (state: PoserState, nextPose: Pose) => {
+  const { dimensions, element, elementStyler } = state.props;
   if (!dimensions.has()) return {};
 
   const prev = dimensions.get() as BoundingBox;
@@ -155,7 +152,18 @@ const implicitlyFlipPose = (state: PoseSetterFactoryProps, nextPose: Pose) => {
   };
 };
 
-export const flipPose = (state: PoseSetterFactoryProps, nextPose: Pose) =>
+export const flipPose = (props: PoserState, nextPose: Pose) =>
   hasPositionalProps(nextPose)
-    ? explicitlyFlipPose(state, nextPose)
-    : implicitlyFlipPose(state, nextPose);
+    ? explicitlyFlipPose(props, nextPose)
+    : implicitlyFlipPose(props, nextPose);
+
+// Prevents the bug where TS errors between "export cannot be named"
+// and import is "declared but unused".
+export {
+  Action,
+  Dimensions,
+  ColdSubscription,
+  DomPopmotionPoser,
+  Poser,
+  Value
+};