+++ packages/popmotion-pose/src/inc/transition-composers.ts
@@ -1,12 +1,12 @@
 import { Transition } from '../types';
 
 type TransitionMap = {
-  [key: string]: Transition
+  [key: string]: Transition;
 };
 
 type TransitionMapFactory = (key: string) => (map: TransitionMap) => Transition;
 
-const createTransitionMap: TransitionMapFactory = (key) => (map) => (props) => {
+const createTransitionMap: TransitionMapFactory = key => map => props => {
   const switchKey = props[key];
   const transition = map[switchKey] || map.default;
   return transition ? transition(props) : false;