+++ packages/pose-core/src/factories/transitions.ts
@@ -1,8 +1,13 @@
 import { PoseMap, Pose, TransitionFactory } from '../types';
+import { invariant } from 'hey-listen';
 
 type DefaultTransitions<A> = Map<string, TransitionFactory<A>>;
 
-const applyDefaultTransition = <A>(pose: Pose<A>, key: string, defaultTransitions: DefaultTransitions<A>): Pose<A> => {
+const applyDefaultTransition = <A>(
+  pose: Pose<A>,
+  key: string,
+  defaultTransitions: DefaultTransitions<A>
+): Pose<A> => {
   return {
     ...pose,
     transition: defaultTransitions.has(key)
@@ -11,15 +16,24 @@ const applyDefaultTransition = <A>(pose: Pose<A>, key: string, defaultTransition
   };
 };
 
-const generateTransitions = <A>(poses: PoseMap<A>, defaultTransitions: DefaultTransitions<A>): PoseMap<A> => {
+const generateTransitions = <A>(
+  poses: PoseMap<A>,
+  defaultTransitions: DefaultTransitions<A>
+): PoseMap<A> => {
   Object.keys(poses).forEach(key => {
     const pose = poses[key];
+
+    invariant(
+      typeof pose === 'object',
+      `Pose '${key}' is of invalid type. All poses should be objects.`
+    );
+
     poses[key] = pose.transition
       ? pose
-      : applyDefaultTransition<A>(pose, key, defaultTransitions)
+      : applyDefaultTransition<A>(pose, key, defaultTransitions);
   });
 
   return poses;
-}
+};
 
 export default generateTransitions;