+++ packages/react-pose/src/components/PoseElement.tsx
@@ -5,18 +5,17 @@ import {
   ChildRegistration,
   CurrentPose,
   PoseContextProps,
-  PoseElementProps,
+  PoseElementInternalProps,
   PopStyle
 } from './PoseElement.types';
-import reactLifecyclePolyfill = require('react-lifecycles-compat');
 
 export const PoseParentContext = createContext({});
 
 type Ref = (ref: Element) => any;
 type RefSetters = {
-  ref?: Ref,
-  innerRef?: Ref,
-  hostRef?: Ref
+  ref?: Ref;
+  innerRef?: Ref;
+  hostRef?: Ref;
 };
 
 const calcPopFromFlowStyle = (el: HTMLElement): PopStyle => {
@@ -40,8 +39,8 @@ const objectToMap = (obj: { [key: string]: any }): Map<string, any> =>
     return map;
   }, new Map());
 
-class PoseElement extends React.PureComponent<PoseElementProps> {
-  props: PoseElementProps;
+class PoseElement extends React.PureComponent<PoseElementInternalProps> {
+  props: PoseElementInternalProps;
   poser: Poser;
   ref: Element;
   styleProps: { [key: string]: any };
@@ -134,7 +133,6 @@ class PoseElement extends React.PureComponent<PoseElementProps> {
     return props;
   }
 
-
   /**
    * We need to get a ref to the underlying DOM element. Styled Components and
    * other libraries use `innerRef`, though this will be swallowed if the
@@ -157,7 +155,7 @@ class PoseElement extends React.PureComponent<PoseElementProps> {
     }
 
     return refs;
-  }
+  };
 
   setRef = (ref: Element) => {
     if (ref instanceof Element || (this.ref && ref === null)) {
@@ -165,7 +163,7 @@ class PoseElement extends React.PureComponent<PoseElementProps> {
       if (innerRef) innerRef(ref);
       this.ref = ref;
     }
-  }
+  };
 
   componentDidMount() {
     if (!this.ref) return;
@@ -178,7 +176,7 @@ class PoseElement extends React.PureComponent<PoseElementProps> {
       values,
       parentValues,
       onDragStart,
-      onDragEnd,
+      onDragEnd
     } = this.props;
     const props: PoserProps = {
       ...poseProps,
@@ -205,11 +203,11 @@ class PoseElement extends React.PureComponent<PoseElementProps> {
     }
   }
 
-  UNSAFE_componentWillUpdate({ pose }: PoseElementProps) {
+  UNSAFE_componentWillUpdate({ pose }: PoseElementInternalProps) {
     if (hasPose(pose, 'flip')) this.poser.measure();
   }
 
-  componentDidUpdate(prevProps: PoseElementProps) {
+  componentDidUpdate(prevProps: PoseElementInternalProps) {
     const { pose } = this.props;
     this.poser.setTransitionProps(this.getSetProps());
     if (pose !== prevProps.pose || pose === 'flip') this.setPose(pose);
@@ -261,7 +259,5 @@ class PoseElement extends React.PureComponent<PoseElementProps> {
     );
   }
 }
-
-reactLifecyclePolyfill(PoseElement);
 
 export { PoseElement };