+++ packages/popmotion-pose/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Pose adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] 2018-05-09
+
+- Migration to `pose-core`
+- Deprecated `transformProps` -> `props` and `setTransitionProps` -> `setProps`
+
 ## [1.2.0] 2018-04-10
 
 ### Added